Adds some more stuff to the tubular belts.
The tubular belts were barely used due to the lack of items that could be placed in them. This will increase the amount of stuff that can be added to them within reason. Everything added has a cylindrical shape or is small enough to be added and give the belt more utility.