Hypospray rework 2.0
The long awaited hypospray rework.
Reworks Hyposprays into MID Apollo Kits.
The MID apollo takes 45u cartridges,
scans a target for it's reagents, displaying it on a toggleable displayer,
And prevents overdoses upon use.
It dispenses in 5, 10, and 15 u doses.
When emagged, it bypasses the overdose check, and injects instantly, similar to current hyposprays.
Adds variants to the MID Apollo known as the 'Hecate' and the 'Ares' Respectably.
Ares is the pre-emagged combat version for traitors, and can swap cartridges like a pistol in a much smoother fashion.
The Hecate is the CMO variant, and the 'original' MID found on the northernlight, which Nanotrasen bases the Apollo copy off of.