Ghosts can occupy borers whenever (#6012)
* Allows ghosts to occupy borer's

About The Pull Request:
Allows ghosts to occupy borer's by clicking on them, checks to see if the borer is already occupied. There is a small bug where eris seems to randomly place the borer around maint, I would fix this but I think it's not a terrible thing even if a bit unimmersive. I'll try if needed.
Why It's Good For The Game:
Borer's as far as I know cannot currently be occupied after a borer breeds.
Changelog:
🆑
add: ghosts can occupy unoccupied borer's
/🆑

* Yes

Yes1

* Removes space

* Adds THE

* Sets outer to false

* Update borer.dm

* Update borer.dm