Elevator bug fixes and cleanup system.
This PR fixes one of the main issues the elevator had: Not deleting IDs.

- The elevator will now check to see if your ID is the captain's spare, and if it is not, it will delete the ID correctly without dropping it like it used to.
- The elevator will now clean up all items inside it (except for the intercom obviously) and put them into a nice little lost and found device outside the elevator. Do mind, the cleanup process is called when the elevator is called DOWN, i.e., when someone presses the call elevator button outside the elevator.
- Additionally did some commenting and make some of the code a little cleaner and easier to read.

This is part 2 of the elevator. Part 3 will come soon, and will include some small easter eggs, an emag option for the elevator and POSSIBLY (if I can figure out how to code it without pulling my hear out) a system that allows players to join via coming down the elevator in order to improve immersion.