Adding Instructional materials
Adding smaller instructional materials that act somewhat like oddities but don't give perks and only guarantee a roll in a stat instead of providing bonus points. Adds them to both the Eris map and Port Oscar. Currently they're hardmapped because spawners are terrible. Also includes a few minor edits to port oscar to allow for their placement.

Unrelated to instructionals, made a small map edit to Port Oscar that adds an APC to the inspectors office and gives it the inspector's office area.