elevator update 3.0
This makes the elevator panel actually do what it was originally intended to do. Ideally it would have a full on detailed interface that looked like an elevator but... yeah no. I'm not that good of a coder. For now though, this will do nicely.

Also cleaned up the code so it's much nicer looking, with given procs that can be used easily in the future if need be.